chore: rename files to maintain directory structure
