+++ src/entities/event.entity.ts
