+++ src/entities/user.entity.ts
